requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
bpemb==0.3.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.12
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
flair==0.8.0.post1
ftfy==6.0.3
future==0.18.2
gdown==3.12.2
gensim==3.8.3
huggingface-hub==0.0.12
hyperopt==0.2.5
idna==3.2
importlib-metadata==3.10.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
Janome==0.4.1
jedi==0.18.0
jieba3k==0.35.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
konoha==4.6.5
langdetect==1.0.9
lxml==4.6.3
matplotlib==3.4.2
matplotlib-inline==0.1.2
mpld3==0.3
networkx==2.5.1
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.21.0
overrides==3.1.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
segtok==1.5.10
sentencepiece==0.1.95
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.1.0
soupsieve==2.2.1
sqlitedict==1.7.0
tabulate==0.8.9
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.10.3
torch==1.7.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
transformers==4.8.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.1
